refactor: unify tilt magnitude and smoothness energy modules with physics parity guards
